+++ test/escompress_test.dart
@@ -1,3 +1,5 @@
+// @dart=2.10
+
 import 'dart:io';
 
 import 'package:test/test.dart';
@@ -8,8 +10,8 @@ import '../bin/es_compress.dart' as escompress;
 void main() {
   const tempFilename = 'helloDart.dart';
   const tempContents = 'hello dart';
-  late final Directory tempDir;
-  late final String inputFile;
+  Directory tempDir;
+  String inputFile;
 
   setUpAll(() {
     tempDir = Directory.systemTemp.createTempSync('es_');